chores: Reduce circular dependency
